refactor(products): extract product card component and use products hook
- Create reusable ProductItem component for both grid and list views
- Replace direct API calls with useProducts hook in FeaturedProducts, Detail and ProductsPage
- Remove duplicate product card logic and consolidate into single component
- Update query keys and parameters in useProducts hook
